делит код на модули и настраивает сборку
