+++ src/components/boardsContainer.js
@@ -1,0 +1,7 @@
+const getBoardsContainer = () => {
+  return (
+    `<section class="films"></section>`
+  );
+};
+
+export {getBoardsContainer};
