+++ src/components/filmsAmountElement.js
@@ -1,0 +1,7 @@
+const getFilmsAmountElement = () => {
+  return (
+    `<p>000 movies inside</p>`
+  );
+};
+
+export {getFilmsAmountElement};
