+++ src/components/mostCommented.js
@@ -1,0 +1,11 @@
+const getMostCommented = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Most commented</h2>
+
+      <div class="films-list__container"></div>
+    </section>`
+  );
+};
+
+export {getMostCommented};
