+++ src/components/showMoreButton.js
@@ -1,0 +1,7 @@
+const getShowMoreButton = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {getShowMoreButton};
